refactored audit.go so that we can have better maintainability, there was duplification of code
